Cleanup stale code
