--- PboExplorer/MainWindow.xaml
@@ -1,144 +1,0 @@
-﻿<Window x:Class="PboExplorer.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:PboExplorer"
-        xmlns:views="clr-namespace:PboExplorer.Modules.Core.Views"
-        xmlns:converters="clr-namespace:PboExplorer.Converters"
-        mc:Ignorable="d"
-        Title="PBO Explorer" Height="450" Width="800">
-    
-    <Window.CommandBindings>
-        <CommandBinding Command="Copy" Executed="CopyToClipboard" CanExecute="CanCopyToClipboard" />
-        <CommandBinding Command="Save" Executed="ExtractSelected" CanExecute="CanExtractSelected" />
-        <CommandBinding Command="Replace" Executed="ReplaceSelected" CanExecute="CanExtractSelected" />
-        <CommandBinding Command="Open" Executed="OpenFile" />
-    </Window.CommandBindings>
-
-    <Window.InputBindings>
-        <KeyBinding Key="C" Modifiers="Control" Command="Copy"/>
-        <KeyBinding Key="S" Modifiers="Control" Command="Save"/>
-        <KeyBinding Key="O" Modifiers="Control" Command="Open"/>
-    </Window.InputBindings>
-
-    <Window.Resources>
-        <converters:TreeItemToIconConverter x:Key="IconConverter"/>
-        
-        <HierarchicalDataTemplate x:Key="PboTreeTemplate" ItemsSource="{Binding Children}">
-            <StackPanel Orientation="Horizontal">
-                <Image Source="{Binding ., Mode=OneWay, Converter={StaticResource IconConverter}}" Margin="0 0 4 0"/>
-                <TextBlock Text="{Binding Name}" />
-            </StackPanel>
-        </HierarchicalDataTemplate>
-        
-        <HierarchicalDataTemplate x:Key="TreeTemplate" ItemsSource="{Binding Children}">
-            <TextBlock Text="{Binding Name}" />
-        </HierarchicalDataTemplate>
-    </Window.Resources>
-    
-    <DockPanel>
-        <Menu DockPanel.Dock="Top">
-            <MenuItem Header="_File">
-                <MenuItem Header="_Open file..." Command="Open" InputGestureText="Ctrl+O" Icon="{StaticResource OpenFileIcon}" />
-                <MenuItem Header="Open all from _directory..." Click="OpenDirectory" Icon="{StaticResource OpenFolderIcon}" />
-                <!--<MenuItem Header="_Close"/>-->
-                <MenuItem Header="Close _all" Click="CloseAll" Icon="{StaticResource CloseAllIcon}" />
-            </MenuItem>
-            <MenuItem Header="_Edit">
-                <MenuItem Header="Replace current file..." Command="Replace" Icon="{StaticResource ReplaceFileIcon}" />
-            </MenuItem>
-            <MenuItem Header="_Extract">
-                <MenuItem Header="_All current PBO files..." Click="ExtractCurrentPBO" Name="ExtractPBO" IsEnabled="False"
-                          Icon="{StaticResource ExtractPBOIcon}"/>
-                <MenuItem Header="_Current file..." Command="Save" InputGestureText="Ctrl+S" Icon="{StaticResource ExtractFileIcon}"/>
-                <MenuItem Header="Current file as _text..." Click="ExtractSelectedAsText" Name="ExtractFileText" IsEnabled="False"
-                          Icon="{StaticResource ExtractTextIcon}" />
-                <MenuItem Header="Current file as _PNG..."  Click="ExtractSelectedAsPNG" Name="ExtractFilePNG" IsEnabled="False"
-                          Icon="{StaticResource ExtractImageIcon}" />
-                <MenuItem Header="Copy to Clipboard" Command="Copy" InputGestureText="Ctrl+C" Icon="{StaticResource CopyIcon}" />
-            </MenuItem>
-            <MenuItem Header="_About" Click="About"/>
-        </Menu>
-        <Grid>
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="1*" />
-                <ColumnDefinition Width="2*" />
-            </Grid.ColumnDefinitions>
-
-            <TabControl  Grid.Column="0">
-                <TabItem Header="PBO Files" AllowDrop="True" Drop="PboFiles_Drop">
-                    <TreeView Name="PboView" ItemTemplate="{StaticResource PboTreeTemplate}" SelectedItemChanged="OnPboViewSelectedItemChanged" VirtualizingStackPanel.IsVirtualizing="True">
-
-                    </TreeView>
-
-                </TabItem>
-
-                <TabItem Header="Config">
-                    <!--<DockPanel>
-                        <Grid DockPanel.Dock="Top">
-                            <TextBox VerticalAlignment="Center" HorizontalAlignment="Stretch" x:Name="SearchConfigTextBox" Margin="5" TextChanged="SearchConfigTextBox_TextChanged" />
-                            <TextBlock IsHitTestVisible="False" Text="Enter Search Term Here" VerticalAlignment="Center" HorizontalAlignment="Left" Margin="10,0,0,0" Foreground="DarkGray">
-                                <TextBlock.Style>
-                                    <Style TargetType="{x:Type TextBlock}">
-                                        <Setter Property="Visibility" Value="Collapsed"/>
-                                        <Style.Triggers>
-                                            <DataTrigger Binding="{Binding Text, ElementName=SearchTermTextBox}" Value="">
-                                                <Setter Property="Visibility" Value="Visible"/>
-                                            </DataTrigger>
-                                        </Style.Triggers>
-                                    </Style>
-                                </TextBlock.Style>
-                            </TextBlock>
-                        </Grid>-->
-                        <TreeView Name="ConfigView" ItemTemplate="{StaticResource TreeTemplate}" SelectedItemChanged="OnConfigViewSelectedItemChanged" VirtualizingStackPanel.IsVirtualizing="True">
-                        </TreeView>
-                    <!--</DockPanel>-->
-                </TabItem>
-                
-                <TabItem Header="Game Data">
-                    <TreeView Name="DataView" ItemTemplate="{StaticResource PboTreeTemplate}" SelectedItemChanged="OnPboViewSelectedItemChanged" VirtualizingStackPanel.IsVirtualizing="True">
-
-                    </TreeView>
-
-                </TabItem>
-            </TabControl>
-
-            <GridSplitter Grid.Column="1" 
-              HorizontalAlignment="Left" 
-              VerticalAlignment="Stretch"
-              Width="5"/>
-
-            <Grid Grid.Column="2" Margin="5 0 0 0">
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="200" />
-                    <RowDefinition />
-                </Grid.RowDefinitions>
-                <DataGrid Name="PropertiesGrid" IsReadOnly="True" AutoGenerateColumns="False" Grid.Row="0">
-                    <DataGrid.Columns>
-                        <DataGridTextColumn Header="Property" Binding="{Binding Name}" Width="1*"></DataGridTextColumn>
-                        <DataGridTextColumn Header="Value" Binding="{Binding Value}" Width="3*"></DataGridTextColumn>
-                    </DataGrid.Columns>
-                </DataGrid>
-                <GridSplitter Grid.Row="1" 
-                      HorizontalAlignment="Stretch" 
-                      VerticalAlignment="Top"
-                      ResizeDirection="Rows"
-                      Height="5"/>
-                <Grid Background="LightGray" Grid.Row="1" Margin="0 5 0 0">
-                    <!--<views:AboutView x:Name="AboutBox"/>-->
-                    <Label x:Name="AboutBox"/>
-                    <TextBox Name="TextPreview" TextWrapping="Wrap" AcceptsReturn="True" VerticalScrollBarVisibility="Visible" Visibility="Hidden" FontFamily="Consolas" IsReadOnly="True" IsReadOnlyCaretVisible="True" />
-                    <views:ImagePreviewView x:Name="ImagePreview" ClipToBounds="True"/>
-                </Grid>
-
-            </Grid>
-            
-        </Grid>
-        
-        
-        
-        
-        
-    </DockPanel>
-</Window>
