--- PboExplorer/PropertyItem.cs
@@ -1,14 +1,0 @@
-﻿namespace PboExplorer
-{
-    public class PropertyItem
-    {
-        public PropertyItem(string name, string value)
-        {
-            Name = name;
-            Value = value;
-        }
-        public string Name { get; }
-
-        public string Value { get; }
-    }
-}
